Add note-saver component for PUTting edited notes and POSTing new notes.
Add actual filtering to NotesList.
Clicking on category or priority will filter notes based on what was clicked.
Add ability to edit notes while on NotesDetails. Note body and status of Task can be edited on any page.
Add ability to add new notes when clicken on 'New note'.
Add methods to PUT and POST notes to NotesService.
Add logic to indicate that the notes were changed only when the note property is different from the initial value of that property (done quite poorly).